career-assistant/requirements.txt
# Core Dependencies
streamlit>=1.32.0
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0

# LLM Providers
groq>=0.5.0
huggingface-hub>=0.19.0

# Data & Analytics
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.17.0
matplotlib>=3.8.0

# Web Dashboard
flask>=2.3.0
flask-cors>=4.0.0
flask-json>=0.3.4

# Database & Storage
sqlalchemy>=2.0.0

# File Handling
pillow>=10.0.0
reportlab>=4.0.0
PyPDF2>=3.0.0
python-docx>=0.8.11
